+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.6.2] - 2020-02-27
+### Added
+- Reports stored on the server
+### Changed
+- Counting set bits in binary string
+### Fixed
+- Publishers' domains in targeting reach
+
 ## [1.6.1] - 2020-02-20
 ### Fixed
 - Response when campaign/site total statistics are empty 
@@ -391,7 +399,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.0]
 
-[Unreleased]: https://github.com/adshares/adserver/compare/v1.6.1...develop
+[Unreleased]: https://github.com/adshares/adserver/compare/v1.6.2...develop
+[1.6.2]: https://github.com/adshares/adserver/compare/v1.6.1...v1.6.2
 [1.6.1]: https://github.com/adshares/adserver/compare/v1.6.0...v1.6.1
 [1.6.0]: https://github.com/adshares/adserver/compare/v1.5.5...v1.6.0
 [1.5.5]: https://github.com/adshares/adserver/compare/v1.5.4...v1.5.5